requirements.txt
streamlit>=1.31.0
pandas>=2.1.4
numpy>=1.26.3
matplotlib>=3.8.2
seaborn>=0.13.1
scikit-learn>=1.4.0
scipy>=1.11.4
google-generativeai>=0.3.2
openpyxl>=3.1.5
protobuf>=5.0